requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
httpx==0.25.2
pydantic[email]==2.5.0
textblob==0.17.1
nltk==3.8.1
wordcloud==1.9.2
pydantic-settings==2.1.0
